make RollNewWord helper, render buttons from map
